requirements.txt
absl-py==2.1.0
agora_token_builder==1.0.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.4.0
altgraph==0.17.4
annotated-types==0.6.0
anthropic==0.21.3
anyio==4.3.0
arabic-reshaper==3.0.0
asgiref==3.4.1
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
Brotli==1.1.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
chardet==5.2.0
charset-normalizer==2.0.4
click==8.1.7
cloudpathlib==0.18.1
cmdstanpy==1.2.4
colorama==0.4.6
comm==0.2.2
confection==0.1.5
constantly==15.1.0
contourpy==1.2.0
cryptography==3.4.7
cssselect2==0.7.0
CT3==3.3.3.post1
curl_cffi==0.6.2
cycler==0.12.1
cymem==2.0.8
daphne==3.0.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==4.0.7
django-allauth==0.63.6
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
dnspython==2.6.1
docopt==0.6.2
duckduckgo_search==5.2.2
eventlet==0.36.1
executing==2.0.1
fastapi==0.112.1
fastjsonschema==2.20.0
fastlogging==1.0.0
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.6.0
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
groq==0.4.2
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h5py==3.11.0
holidays==0.55
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hyperlink==21.0.0
idna==3.2
importlib_resources==6.4.4
incremental==21.3.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupytext==1.16.4
keras==3.5.0
keras-models==0.0.7
keybert==0.8.4
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mime==0.1.0
mistletoe==1.3.0
mistralai==0.0.8
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
namex==0.0.8
narwhals==1.5.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netlify-py==0.1.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.1.1
ollama==0.1.8
openai==1.16.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
oscrypto==1.3.0
outcome==1.3.0.post0
p2j==1.3.2
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
patsy==0.5.6
pdfminer.six==20231228
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.1
playwright==1.42.0
pluggy==1.4.0
preshed==3.0.9
prompt-toolkit==3.0.43
prophet==1.1.5
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.1
pyee==11.0.1
Pygments==2.17.2
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.4.0
pyOpenSSL==20.0.1
pyparsing==3.1.2
pypdf==4.1.0
pypng==0.20220715.0
PySocks==1.7.1
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-engineio==4.9.0
python-slugify==8.0.4
python-socketio==5.11.2
python3-openid==3.2.0
pytz==2021.1
PyWavelets==1.6.0
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
rapidfuzz==3.9.5
referencing==0.35.1
regex==2024.7.24
reportlab==4.0.9
requests==2.32.3
requests-oauthlib==1.3.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selenium==4.23.1
sentence-transformers==2.6.1
service-identity==21.1.0
setuptools==72.1.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
smtpmail==1.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
sqlmodel==0.0.16
sqlparse==0.4.2
srsly==2.4.8
stack-data==0.6.3
stanio==0.5.1
starlette==0.38.2
statsmodels==0.14.2
streamlit==1.36.0
streamlit-option-menu==0.3.13
svglib==1.5.1
sympy==1.12
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.39.3
trio==0.26.0
trio-websocket==0.11.1
Twisted==22.4.0
twisted-iocpsupport==1.0.4
txaio==21.2.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.6
uvicorn==0.30.6
virtualenv==20.26.1
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
xhtml2pdf==0.2.15
yarg==0.1.9
yarl==1.9.4
zope.event==5.0
zope.interface==5.4.0
PyPDF2